+++ src/components/Register.jsx
@@ -116,7 +116,7 @@ export default function Register() {
             </Col>
           </Row>
 
-          <div style={{display: 'flex', flexDirection: 'row', justifyContent: 'end', margin: '2rem 0 0 0'}}>
+          <div style={{display: 'flex', flexDirection: 'row', justifyContent: 'end', marginTop: '2rem'}}>
             <Button 
               onClick={register} 
               color='purple'
